pyproject.toml
[tool.poetry]
name = "java-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^14.0.0"
mcstatus = "^12.0.1"
docker = "^7.1.0"
python-dotenv = "^1.1.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
